Generate 1.0.12.0@147667

+++ Source/BF_FrameworkGame/Public/IGS_PlayerStateData.h
@@ -15,7 +15,7 @@ public:
     EIGS_StorePlatform StorePlatform;
 
     UPROPERTY(BlueprintReadWrite)
-    int32 HordeModeTokens;
+    uint8 HordeModeTokens;
 
     FIGS_PlayerStateData();
 };
